+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.5.0]
+## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.7.0]
+
+### Added
+
+- bincount2D moved from ibllib.processing to iblutil.numerical
+
+## [1.6.0]
+
+### Added
+
+- numerical.rcoeff function that computes pairwise Pearson correlation coefficients for matrices
+
+## [1.5.0]
 
 ### Added